extract usage into utils
