+++ packages/server/lib/controllers/v1/plans/usage/getUsage.ts
@@ -1,12 +1,9 @@
-import { getAccountUsageTracker } from '@nangohq/account-usage';
-import { connectionService } from '@nangohq/shared';
 import { requireEmptyQuery, zodErrorToHTTP } from '@nangohq/utils';
 
 import { asyncWrapper } from '../../../../utils/asyncWrapper.js';
+import { getAccountMetricsUsage } from '../../../../utils/usage.js';
 
 import type { GetUsage } from '@nangohq/types';
-
-const accountUsageTracker = await getAccountUsageTracker();
 
 export const getUsage = asyncWrapper<GetUsage>(async (req, res) => {
     const emptyQuery = requireEmptyQuery(req, { withEnv: true });
@@ -21,34 +18,9 @@ export const getUsage = asyncWrapper<GetUsage>(async (req, res) => {
         return;
     }
 
+    const usage = await getAccountMetricsUsage(account, plan);
+
     res.status(200).send({
-        data: [
-            {
-                metric: 'connections',
-                label: 'Connections',
-                usage: await connectionService.countByAccountId(account.id),
-                limit: plan.connections_max
-            },
-            {
-                metric: 'actions',
-                label: 'Actions',
-                usage:
-                    (await accountUsageTracker.getUsage({
-                        accountId: account.id,
-                        metric: 'actions'
-                    })) ?? 0,
-                limit: plan.monthly_actions_max
-            },
-            {
-                metric: 'active_records',
-                label: 'Synced Records',
-                usage:
-                    (await accountUsageTracker.getUsage({
-                        accountId: account.id,
-                        metric: 'active_records'
-                    })) ?? 0,
-                limit: plan.monthly_active_records_max
-            }
-        ]
+        data: usage
     });
 });
