+++ packages/types/lib/usage/dto.ts
@@ -11,3 +11,10 @@ export interface UpdateAccountUsageDto {
     actions?: DBAccountUsage['actions'];
     activeRecords?: DBAccountUsage['active_records'];
 }
+
+export interface MetricUsage {
+    metric: string;
+    label: string;
+    usage: number;
+    limit: number | null;
+}
